Check whether a data frame contains all the required variable

Checks whether a data frame contains all the required variables. Matching works via variable names, and used to support individual functions in the package. Not used directly.

Usage

check_inputs(input, requirements, return = "stop")

Arguments

input

Pass a data frame for checking

requirements

A character vector specifying the required variable names

return

A character string specifying what to return. The default value is "stop". Also accepts "names" and "warning".

Value

The default behaviour is to return an error message, informing the user what variables are not included. When return is set to "names", a character vector containing the unmatched variable names is returned.

See also

Examples

# Return error message
if (FALSE) { # \dontrun{
check_inputs(iris, c("Sepal.Length", "mpg"))
} # }

#' # Return warning message
check_inputs(iris, c("Sepal.Length", "mpg"), return = "warning")
#> Warning: The following variables are not included in the input data frame:
#>  mpg

# Return variable names
check_inputs(iris, c("Sepal.Length", "Sepal.Width", "RandomVariable"), return = "names")
#> [1] "RandomVariable"