Create interactive tables in HTML with 'download' buttons.

See https://martinctc.github.io/blog/vignette-downloadable-tables-in-rmarkdown-with-the-dt-package/ for more.

Usage

create_dt(x, rounding = 1, freeze = 2, percent = FALSE)

Arguments

x

Data frame to be passed through.

rounding

Numeric vector to specify the number of decimal points to display

freeze

Number of columns from the left to 'freeze'. Defaults to 2, which includes the row number column.

percent

Logical value specifying whether to display numeric columns as percentages.

Value

Returns an HTML widget displaying rectangular data.

Details

This is exported from wpa::create_dt().

See also

Examples

output <- hrvar_count(pq_data, return = "table")
create_dt(output)