Extract HR attribute variables

This function uses a combination of variable class, number of unique values, and regular expression matching to extract HR / organisational attributes from a data frame.

Usage

extract_hr(data, max_unique = 50, exclude_constants = TRUE, return = "names")

Arguments

data

A data frame to be passed through.

max_unique

A numeric value representing the maximum number of unique values to accept for an HR attribute. Defaults to 50.

exclude_constants

Logical value to specify whether single-value HR attributes are to be excluded. Defaults to TRUE.

return

String specifying what to return. This must be one of the following strings:

"names"

"vars"

See Value for more information.

Value

A different output is returned depending on the value passed to the return

argument:

"names": character vector identifying all the names of HR variables present in the data.

"vars": data frame containing all the columns of HR variables present in the data.

See also

Examples

pq_data %>% extract_hr(return = "names")
#> [1] "LevelDesignation"    "SupervisorIndicator" "Organization"       
#> [4] "FunctionType"        "WeekendDays"        

pq_data %>% extract_hr(return = "vars")
#> # A tibble: 5,593 × 5
#>    LevelDesignation SupervisorIndicator    Organization        Functio…¹ Weeke…²
#>    <chr>            <chr>                  <chr>               <chr>     <chr>  
#>  1 Junior IC        Individual Contributor Product             G_and_A   [SUNDA…
#>  2 Manager          Manager                Sales and Marketing G_and_A   [SUNDA…
#>  3 Support          Individual Contributor Finance             Engineer… [SUNDA…
#>  4 Support          Individual Contributor Finance             Sales     [SUNDA…
#>  5 Director         Manager+               Sales and Marketing Customer… [SUNDA…
#>  6 Support          Individual Contributor Product             IT        [SUNDA…
#>  7 Director         Manager+               Sales and Marketing Analytics [SUNDA…
#>  8 Support          Individual Contributor Sales and Marketing IT        [SUNDA…
#>  9 Support          Individual Contributor Finance             IT        [SUNDA…
#> 10 Support          Individual Contributor HR                  Engineer… [SUNDA…
#> # … with 5,583 more rows, and abbreviated variable names ¹​FunctionType,
#> #   ²​WeekendDays