Identify Non-Knowledge workers in a Person Query using Collaboration Hours

This function scans a standard query output to identify employees with consistently low collaboration signals. Returns the % of non-knowledge workers identified by Organization, and optionally an edited data frame with non-knowledge workers removed, or the full data frame with the kw/nkw flag added.

Usage

identify_nkw(data, collab_threshold = 5, return = "data_summary")

Arguments

data

A Standard Person Query dataset in the form of a data frame.

collab_threshold

Positive numeric value representing the collaboration hours threshold that should be exceeded as an average for the entire analysis period for the employee to be categorized as a knowledge worker ("kw"). Default is set to 5 collaboration hours. Any versions after v1.4.3, this uses a "greater than or equal to" logic (>=), in which case persons with exactly 5 collaboration hours will pass.

return

String specifying what to return. This must be one of the following strings:

"text"

"data_with_flag"

"data_clean"

"data_summary"

See Value for more information.

Value

A different output is returned depending on the value passed to the return

argument:

"text": string. Returns a diagnostic message.

"data_with_flag": data frame. Original input data with an additional column containing the kw/nkw flag.

"data_clean": data frame. Data frame with non-knowledge workers excluded.

"data_summary": data frame. A summary table by organization listing the number and % of non-knowledge workers.

See also